Правки по замечаниям: 1. Убрал лишнбб абстракцию ClientRepository 2. Добавил передачу зависимостей через конструктор 3. Убрал несекьюрные названия методов 4. Добавил миграции
