--- hw12-webserver/src/main/java/db/repository/ClientRepository.java
@@ -1,10 +1,0 @@
-package db.repository;
-
-import db.crm.model.Client;
-
-import java.util.List;
-
-public interface ClientRepository {
-    Client saveClientInDb(Client client);
-    List<Client> getAllClientFromDb();
-}
