+++ hw12-webserver/src/main/java/webserver/servlet/ClientsServlet.java
@@ -2,12 +2,11 @@ package webserver.servlet;
 
 import db.crm.model.Address;
 import db.crm.model.Client;
+import db.crm.service.DBServiceClient;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import db.repository.ClientRepository;
-import db.repository.ClientRepositoryImpl;
 import webserver.services.TemplateProcessor;
 
 import java.io.IOException;
@@ -20,18 +19,18 @@ public class ClientsServlet extends HttpServlet {
 
     private static final String CLIENTS_PAGE_TEMPLATE = "clients.html";
 
-    private final ClientRepository clientRepository;
     private final TemplateProcessor templateProcessor;
+    private final DBServiceClient dbServiceClient;
 
-    public ClientsServlet(TemplateProcessor templateProcessor) {
+    public ClientsServlet(TemplateProcessor templateProcessor, DBServiceClient dbServiceClient) {
         this.templateProcessor = templateProcessor;
-        this.clientRepository = new ClientRepositoryImpl();
+        this.dbServiceClient = dbServiceClient;
     }
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         Map<String, Object> paramsMap = new HashMap<>();
-        paramsMap.put("clients", clientRepository.getAllClientFromDb());
+        paramsMap.put("clients", dbServiceClient.findAll());
 
         resp.setContentType("text/html");
         resp.getWriter().println(templateProcessor.getPage(CLIENTS_PAGE_TEMPLATE, paramsMap));
@@ -44,7 +43,7 @@ public class ClientsServlet extends HttpServlet {
         String name = req.getParameter("name");
         String address = req.getParameter("address");
 
-        Client client = clientRepository.saveClientInDb(new Client(null, name, new Address(null, address), new ArrayList<>()));
+        Client client = dbServiceClient.saveClient(new Client(null, name, new Address(null, address), new ArrayList<>()));
         log.info("Added new client: {}", client);
 
         resp.sendRedirect(req.getContextPath() + "/clients");
