+++ hw12-webserver/src/main/resources/hibernate.cfg.xml
@@ -12,11 +12,11 @@
     <property name="hibernate.connection.username">usr</property>
     <property name="hibernate.connection.password">pwd</property>
 
-    <property name="hibernate.show_sql">false</property>
+    <property name="hibernate.show_sql">true</property>
     <property name="hibernate.format_sql">false</property>
     <property name="hibernate.generate_statistics">true</property>
 
-    <property name="hibernate.hbm2ddl.auto">create</property>
+    <property name="hibernate.hbm2ddl.auto">none</property>
     <property name="hibernate.enable_lazy_load_no_trans">false</property>
 
   </session-factory>
